cosmetic fix to a few errors

+++ front-end/WithIt/MasterViewController.m
@@ -201,11 +201,11 @@
     switch (section){
         case 0:
             numRows = [self.dataController.masterPollsList count];
-            NSLog(@"Number of friends' polls: %d.", numRows);
+            NSLog(@"Number of friends' polls: %lu.", (unsigned long)numRows);
             break;
         case 1:
             numRows = [self.dataController.masterPollsCreatedList count];
-            NSLog(@"Number of created polls: %d.", numRows);
+            NSLog(@"Number of created polls: %lu.", (unsigned long)numRows);
             break;
     }
     return numRows;
